Refactor RetryRepository and RealmRetryOperation to use RetryFailure
- Introduced `RetryFailure` in the `model` package containing only necessary fields (itemId, message, httpCode) to decouple the repository and model packages from the services package.
- Updated `RealmRetryOperation.createFromRetryFailure` to accept `RetryFailure` instead of `UploadError`.
- Updated `RetryRepository` and `RetryRepositoryImpl` enqueue and updateAttempt methods to accept `RetryFailure`.
- Removed `org.ole.planet.myplanet.services.upload.UploadError` imports from `RetryRepository.kt`, `RetryRepositoryImpl.kt`, and `RealmRetryOperation.kt`.
- Updated `RetryQueue.kt` to map `UploadError` to `RetryFailure` before passing it to the repository layer.
- Updated `RetryRepositoryImplTest.kt` to use `RetryFailure` for test objects and verify correct mapping behavior.
- Updated `RetryQueueTest.kt` to verify correct `RetryFailure` mappings.
