+++ app/src/main/java/org/ole/planet/myplanet/model/RetryFailure.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.repository
+package org.ole.planet.myplanet.model
 
 data class RetryFailure(
     val itemId: String,
