+++ app/src/main/java/org/ole/planet/myplanet/repository/RetryRepository.kt
@@ -1,6 +1,7 @@
 package org.ole.planet.myplanet.repository
 
 import org.ole.planet.myplanet.model.RealmRetryOperation
+import org.ole.planet.myplanet.model.RetryFailure
 
 interface RetryRepository {
     suspend fun enqueue(
